+++ tests/test_operations.py
@@ -1,9 +1,11 @@
 from invisibleroads_macros_disk import (
     TemporaryStorage,
+    make_enumerated_folder,
+    make_random_folder,
     make_unique_folder,
     remove_folder,
     remove_path)
-from os.path import exists
+from os.path import basename, exists
 from tempfile import mkstemp
 
 
@@ -11,6 +13,16 @@ def test_temporary_storage():
     with TemporaryStorage() as storage:
         assert exists(storage.folder)
     assert not exists(storage.folder)
+
+
+def test_make_enumerated_folder(tmpdir):
+    folder = make_enumerated_folder(tmpdir)
+    assert basename(folder) == '1'
+
+
+def test_make_random_folder(tmpdir):
+    folder = make_random_folder(tmpdir, 10)
+    assert len(basename(folder)) == 10
 
 
 def test_remove_folder():