+++ CHANGES.md
@@ -2,3 +2,4 @@
 - Restore TemporaryStorage, make_folder, make_unique_folder
 - Restore remove_folder, remove_path
 - Restore check_absolute_path, get_absolute_path, get_real_path
+- Restore make_enumerated_folder, make_random_folder